:sparkles: (projects) set the initial productibity as float to consider the already started project

+++ pages/api/projects/index.ts
@@ -23,7 +23,7 @@ export default withSentry(async (req: NextApiRequest, res: NextApiResponse) => {
       data: {
         name: req.body.name,
         unit: unit.toLowerCase() === "ticket" ? "TICKET" : "POINT",
-        productivity: parseInt(productivity),
+        productivity: parseFloat(productivity),
         endAt: parseGMTMidnight(endDate),
         startAt: parseGMTMidnight(startDate),
         team: {
